+++ sw.js
@@ -1,4 +1,4 @@
-const CACHE_NAME = 'budgeter-v40';
+const CACHE_NAME = 'budgeter-v41';
 
 const FILES_TO_CACHE = [
   './',